Support AVX/2/512 instructions
Use compiler flags to compile the core piece for the different instruction sets. Use CPUID to determine the platform and dynamic switch at runtime.
