+++ Core/CMakeLists.txt
@@ -11,7 +11,9 @@ file(GLOB SRC "*.cpp" "radio/*.cpp" "pffft/*.cpp" ${ARCH_SRC})
 
 include_directories("." "CyAPI/" "pffft/")
 if (MSVC)
-
+    set_source_files_properties(fft_mt_r2iq_avx.cpp PROPERTIES COMPILE_FLAGS /arch:AVX)
+    set_source_files_properties(fft_mt_r2iq_avx2.cpp PROPERTIES COMPILE_FLAGS /arch:AVX2)
+    set_source_files_properties(fft_mt_r2iq_avx512.cpp PROPERTIES COMPILE_FLAGS /arch:AVX512)
 else()
     include_directories(${LIBUSB_INCLUDE_DIRS})
     include_directories(${LIBFFTW_INCLUDE_DIRS})
